# Conflicts:
#	src/SampleWebApp/Migrations/20230808104126_InitialCreate.Designer.cs
#	src/SampleWebApp/Migrations/20230808104126_InitialCreate.cs
#	src/SampleWebApp/Migrations/20230809091528_InitialCreate.Designer.cs
#	src/SampleWebApp/Migrations/20230809091528_InitialCreate.cs
#	src/SampleWebApp/Migrations/20230809092038_InitialCreate.Designer.cs
#	src/SampleWebApp/Migrations/20230809092038_InitialCreate.cs
#	src/SampleWebApp/Migrations/SampleWebAppDbContextModelSnapshot.cs

+++ src/SampleWebApp/SeedData/Models/CountryModel.cs
@@ -11,7 +11,7 @@
         public string Name { get; set; } = string.Empty;
         public string FormalName { get; set; } = string.Empty;
         public int NumericCode { get; set; }
-        public decimal Population { get; set; }
+        public double Population { get; set; }
         public string TopLevelDomains { get; set; } = string.Empty;
         public string GeoRegion { get; set; } = string.Empty;
         public string GeoSubRegion { get; set; } = string.Empty;
